requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-WTF
psycopg2-binary
python-dotenv
authlib
requests
rank_bm25
gunicorn
